Separate Dev and Prod docker environments
Changing python image from alpine to slim-stretch because psycopg is incompatible with alpine.

Postgres DB with persistence.

In the future, Postgres and Redis images should be removed from the compose app and shifted to separate containers or cloud hosted instances to enable multiple containers to use the same DB and Cache.